% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{combinaison}
\alias{combinaison}
\title{Combinaison}
\usage{
combinaison(k, n)
}
\arguments{
\item{k}{An integer representing the number of elements to choose}

\item{n}{An integer representing the total number of elements}
}
\value{
A numeric value corresponding to the binomial coefficient "k choose n"
}
\description{
This function calculates the binomial coefficient "k choose n",
which represents the number of ways to choose k elements from a
total of n elements
}
\examples{
combinaison(2, 5) #returns 10
}
