% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{f}
\alias{f}
\title{Evaluate a mathematical function at a given value}
\usage{
f(func, x)
}
\arguments{
\item{func}{A string representing a mathematical function}

\item{x}{The value at which to evaluate the function}
}
\value{
A numeric value representing the result of evaluating the function at the given value
}
\description{
The function takes a mathematical function represented as a string
and evaluates it at a specified value. It replaces all the occurences
of 'x' and calculates the value of the function
}
\examples{
f("exp(x)", 5)  #Returns the value of exp(5)
}
