% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Newton-Raphson}
\usage{
newton_raphson(func, x0, erreur = 1e-05)
}
\arguments{
\item{func}{A string representing a mathematical function}

\item{x0}{The initial guess for the root}

\item{erreur}{The desired level of accuracy}
}
\value{
A numeric value representing the approximate root of the function using the Newton-Raphson method
}
\description{
Approximate the root of a function using the Newton-Raphson method
}
\examples{
newton_raphson('x^2 - 3*x - 4', 6) #4.000001
}
