% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numa-package.R
\docType{package}
\name{numa-package}
\alias{numa}
\alias{numa-package}
\title{numa: package developed to facilitate numerical analysis and data manipulation in the field of scientific computing}
\description{
This package provides a collection of powerful and efficient functions that enable solving complex mathematical problems and performing precise numerical calculations.
}
\author{
\strong{Maintainer}: Idriss Wendeu \email{idrisswendeuk@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
