% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polyadd}
\alias{polyadd}
\title{Addition of polynomials}
\usage{
polyadd(p, q)
}
\arguments{
\item{p}{A coefficient vector representing the first polynomial}

\item{q}{A coefficient vector representing the second polynomial}
}
\value{
A coefficient vector representing the resulting polynomial from the addition
}
\description{
Performs the addition of two polynomials represented by coefficient vectors
}
\examples{
polyadd(c(1, 0, -1), c(1, 0, 3, 5)) #output: c(1, 1, 3, 4)
polyprint(polyadd(c(1, 2, 3), c(4, 5, 6)))
}
