% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polymul}
\alias{polymul}
\title{Multiplication of two polynomials}
\usage{
polymul(p, q)
}
\arguments{
\item{p}{A coefficient vector representing the first polynomial}

\item{q}{A coefficient vector representing the second polynomial}
}
\value{
A coefficient vector representing the resulting polynomial from the multiplication
}
\description{
performs the multiplication of two polynomials represented by coefficient vectors
}
\examples{
polymul(c(1, 0, -1), c(3, 0, 1, 5))
polyprint(polymul(c(1, 3), c(4, 5)))
}
