% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polypow}
\alias{polypow}
\title{Raises a polynomial to a given power}
\usage{
polypow(p, n = 1)
}
\arguments{
\item{p}{A coefficient vector representing the first polynomial}

\item{n}{A coefficient vector representing the first polynomial}
}
\value{
A coefficient vector representing the resulting polynomial from the pow
}
\description{
Raises a polynomial to a given power
}
\examples{
polypow(c(1, 0, -1), 3)
polyprint(polypow(c(1, 1), 2))
}
