% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylor}
\alias{taylor}
\title{Taylor}
\usage{
taylor(func, x = 0, n = 0, a = 0, pol = FALSE)
}
\arguments{
\item{func}{A string representing a mathematical function or a polynomial}

\item{x}{The value at which to evaluate the Taylor series}

\item{n}{An integer specifying the degree of the Taylor series}

\item{a}{The point at which to calculate the Taylor series}

\item{pol}{A logical value indicating whether  to return the Taylor polynomial}
}
\value{
if \code{pol} == FALSE, a numeric value representing the approximate value of the function

if \code{pol} == TRUE, a polynomial object representing the Taylor polynomial
}
\description{
Approximates a function using Taylor series expansion
}
\examples{
taylor('sin(x)', x = 1, n = 3, a = 0) #0.8283376
}
