% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{trapezoidal_rule}
\alias{trapezoidal_rule}
\title{Trapezoidal rule}
\usage{
trapezoidal_rule(func, a, b, n = 10)
}
\arguments{
\item{func}{A string representing a mathematical function}

\item{a}{The lower limit of integration}

\item{b}{The upper limit of integration}

\item{n}{An integer representing the number of subintervals}
}
\value{
A numeric value representing the approximate value of the definite integral using the trapezoidal rule
}
\description{
Approximates the definite integral of a function using the trapezoidal rule
}
\examples{
trapezoidal_rule('log(x)', 1, 2, 1000)
}
